+++ packages/tools/src/_cleanup/cleanup-prettierrc.ts
@@ -1,0 +1,17 @@
+export default {
+  printWidth: 80,
+  tabWidth: 2,
+  useTabs: true,
+  semi: true,
+  singleQuote: true,
+  jsxSingleQuote: false,
+  trailingComma: 'all',
+  bracketSpacing: true,
+  jsxBracketSameLine: false,
+  arrowParens: 'always',
+  requirePragma: false,
+  insertPragma: false,
+  proseWrap: 'preserve',
+  htmlWhitespaceSensitivity: 'css',
+  endOfLine: 'lf',
+};